Set up new dependencies to work MapStruct with SpringTest. Impl first test, fulfilled services logic.
Now accessing and converting data from nbp api is working fine.

+++ src/main/java/pl/dashboard/nbp/features/dataprovider/dto/NbpExternalApiData.java
@@ -25,8 +25,8 @@ public class NbpExternalApiData {
     private String code;
 
     @JsonProperty("bid")
-    private Long bid;
+    private Float bid;
 
     @JsonProperty("ask")
-    private Long ask;
+    private Float ask;
 }
